backend/pyproject.toml
[tool.poetry]
name = "podcharts-backend"
version = "0.1.0"
description = "FastAPI backend for PodCharts"
authors = ["PodCharts <dev@podcharts.xyz>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
python-dotenv = "^1.0.1"
psycopg = {version = "^3.2.1", extras = ["binary"]}
psycopg-pool = "^3.2.1"
redis = "^5.0.8"
httpx = "^0.27.2"
supabase = "^2.0.0"
stripe = "^10.0.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.9"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.8"
mypy = "^1.11.1"
types-redis = "^4.6.0.20241004"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"

